+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `border.active` token to help with conveying the status of a component.
+- Added `underline` variant for `Link` component.
 
 ## Changed